Next.js frontend working
