+++ .gitignore
@@ -6,17 +6,18 @@
 /.pnp
 .pnp.js
 
+# next.js
+/.next/
+/out/
+
 # uploads
-/server/uploads
+/uploads
 
 # testing
 /coverage
 
 # production
 /build
-
-# Server/backend
-server/secrets.js
 
 # TypeScript v1 declaration files
 typings/
@@ -48,17 +49,14 @@ typings/
 # dotenv environment variables file
 .env
 .env.test
-/server/.env
-/server/.env.test
-/client/.env
-/client/.env.test
-
-# misc
-.DS_Store
+/src/.env
 .env.local
 .env.development.local
 .env.test.local
 .env.production.local
+
+# misc
+.DS_Store
 
 npm-debug.log*
 yarn-debug.log*
