--- client/src/pages/Home.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Home = () => {
-  return <>Home</>;
-};
-
-export default Home;
