--- client/src/pages/NotFound.js
@@ -1,5 +1,0 @@
-import React from "react";
-
-const NotFound = () => <>404</>;
-
-export default NotFound;
