+++ package.json
@@ -1,32 +1,32 @@
 {
   "name": "shop",
   "version": "1.0.0",
-  "description": "An e-commerce website built with the MERN stack.",
-  "main": "server.js",
+  "description": "An e-commerce website built with Next.js",
   "scripts": {
-    "start": "concurrently \"npm run client\" \"npm run server\"",
-    "start:nowatch": "concurrently \"npm run client\" \"npm run server:nowatch\"",
-    "server": "cd server && nodemon server.js",
-    "server:nowatch": "cd server && node server.js",
-    "client": "cd client && npm start"
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start"
   },
-  "keywords": [],
   "author": "Forrest Wilkins",
   "license": "GPL-3.0",
   "dependencies": {
-    "bcrypt": "^5.0.0",
-    "body-parser": "^1.19.0",
-    "concurrently": "^5.3.0",
-    "cors": "^2.8.5",
-    "dotenv": "^8.2.0",
-    "express": "^4.17.1",
-    "is-empty": "^1.2.0",
-    "jsonwebtoken": "^8.5.1",
-    "mongoose": "^5.10.18",
-    "multer": "^1.4.2",
-    "passport": "^0.4.1",
-    "passport-jwt": "^4.0.0",
-    "stripe": "^8.129.0",
-    "validator": "^13.5.1"
+    "@fortawesome/fontawesome-svg-core": "^1.2.32",
+    "@fortawesome/free-solid-svg-icons": "^5.15.1",
+    "@fortawesome/react-fontawesome": "^0.1.14",
+    "@material-ui/core": "^4.11.2",
+    "@material-ui/icons": "^4.11.2",
+    "@stripe/react-stripe-js": "^1.1.2",
+    "@stripe/stripe-js": "^1.11.0",
+    "axios": "^0.21.1",
+    "bootstrap": "^4.5.3",
+    "classnames": "^2.2.6",
+    "jwt-decode": "^3.1.2",
+    "next": "10.0.4",
+    "node-sass": "^5.0.0",
+    "prop-types": "^15.7.2",
+    "react": "17.0.1",
+    "react-bootstrap": "^1.4.0",
+    "react-dom": "17.0.1",
+    "zustand": "^3.2.0"
   }
 }
