--- server/validation/validateLogin.js
@@ -1,32 +1,0 @@
-const validator = require("validator");
-const isEmpty = require("is-empty");
-
-module.exports = function validateLoginInput(data) {
-  let errors = {};
-
-  /* eslint-disable no-param-reassign */
-  data.email = !isEmpty(data.email) ? data.email : "";
-  data.password = !isEmpty(data.password) ? data.password : "";
-  /* eslint-enable no-param-reassign */
-
-  if (!validator.isEmail(data.email)) {
-    errors.email = "Email is invalid";
-  }
-
-  if (validator.isEmpty(data.email)) {
-    errors.email = "Email is required";
-  }
-
-  if (!validator.isLength(data.password, { min: 4, max: 30 })) {
-    errors.password = "Password must have 4 and 30 chars";
-  }
-
-  if (validator.isEmpty(data.password)) {
-    errors.password = "Password is required";
-  }
-
-  return {
-    errors,
-    isValid: isEmpty(errors),
-  };
-};
