+++ src/components/_App/HeadContent.js
@@ -1,0 +1,13 @@
+const HeadContent = () => (
+  <>
+    <meta charSet="utf-8" />
+    <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
+    <meta
+      name="viewport"
+      content="minimum-scale=1, initial-scale=1, width=device-width"
+    />
+    <meta name="theme-color" content="#000000" />
+  </>
+);
+
+export default HeadContent;
