+++ src/models/image.model.js
