+++ src/pages/index.js
@@ -1,0 +1,5 @@
+const Home = () => {
+  return <>Home</>;
+};
+
+export default Home;
