+++ src/utils/auth.js
@@ -1,0 +1,11 @@
+import axios from "axios";
+
+const setAuthToken = (token) => {
+  if (token) {
+    axios.defaults.headers.common.Authorization = token;
+  } else {
+    delete axios.defaults.headers.common.Authorization;
+  }
+};
+
+export { setAuthToken };
