+++ src/utils/formatDate.js
@@ -1,0 +1,5 @@
+const formatDate = date => {
+  return new Date(date).toLocaleDateString("en-US");
+};
+
+export default formatDate;
